requirements.txt
Django~=3.1.13
django-material-admin
djangorestframework
markdown
django-filter